Added DTOs for Products and Orders. Refactored controllers and services

+++ Inventory/Services/ProductService.cs
@@ -1,4 +1,5 @@
 ﻿using Inventory.Api.Domain.Entities;
+using Inventory.Api.DTOs.Products;
 using Inventory.Api.Persistence;
 using Inventory.Api.Services.Interfaces;
 using Microsoft.EntityFrameworkCore;
@@ -14,38 +15,67 @@ namespace Inventory.Api.Services
             _context = context;
         }
 
-        public async Task<Product> CreateAsync(Product product)
+        public async Task<Product> CreateAsync(Guid userId, CreateProductRequest request)
         {
+            var product = new Product
+            {
+                Name = request.Name,
+                Price = request.Price,
+                StockQuantity = request.StockQuantity,
+                UserId = userId,
+            };
+
             _context.Products.Add(product);
             await _context.SaveChangesAsync();
             return product;
         }
 
-        public async Task<List<Product>> GetAllAsync()
+        public async Task<IEnumerable<Product>> GetAllAsync(Guid userId)
         {
-            return await _context.Products.ToListAsync();
+            return await _context.Products
+                .Where(p => p.UserId == userId)
+                .ToListAsync();
         }
 
-        public async Task<List<Product>> GetProductsByUserAsync(Guid userId)
+        public async Task<IEnumerable<Product>> GetProductsByUserAsync(Guid userId)
         {
             return await _context.Products
                                  .Where(p => p.UserId == userId)
                                  .ToListAsync();
         }
 
-        public async Task<Product?> GetByIdAsync(Guid id)
+        public async Task<Product?> GetByIdAsync(Guid userId, Guid productId)
         {
-            return await _context.Products.FindAsync(id);
+            return await _context.Products
+                .FirstOrDefaultAsync(p => p.Id == productId && p.UserId == userId);
         }
 
-        public async Task UpdateAsync(Product product)
+        public async Task UpdateAsync(Guid userId, Guid productId, UpdateProductRequest request)
         {
-            _context.Products.Update(product);
+            var product = await _context.Products
+                .FirstOrDefaultAsync(p => p.Id == productId && p.UserId == userId);
+
+            if (product is null)
+            {
+                throw new KeyNotFoundException("Product not found");
+            }
+
+            product.Name = request.Name;
+            product.Price = request.Price;
+            product.StockQuantity = request.StockQuantity;
             await _context.SaveChangesAsync();
         }
 
-        public async Task DeleteAsync(Product product)
+        public async Task DeleteAsync(Guid userId, Guid productId)
         {
+            var product = await _context.Products
+                .FirstOrDefaultAsync(p => p.Id == productId && p.UserId ==  userId);
+
+            if (product is null)
+            {
+                throw new KeyNotFoundException("Product not found");
+            }
+
             _context.Products.Remove(product);
             await _context.SaveChangesAsync();
         }
